added support for Non-Circular Minimap shapes using the GetMinimapShape() function as defined on WoW Wiki http://www.wowwiki.com/GetMinimapShape
Note: Currently there is a bug in the WoW client with non-circular Minimaps and using the Rotate Minimap UI option which causes the terrain texture to not stay within the expected boarders of the a custom shape mask texture.  I have coded this support for non-circular Minimaps with the assumption in mind that the client will be updated to NOT rotate the Mask Texture when it rotates the Minimap.